Vectorized query, returning only answers in a data frame

Vectorized query, returning only answers in a data frame

bulk_query(entities, type = 1, edns_client_subnet = "0.0.0.0/0")

Arguments

entities

character vector of entities to query

type

RR type can be represented as a number in [1, 65535] or canonical string (A, aaaa, etc). More information on RR types can be found http://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#dns-parameters-4.

edns_client_subnet

The edns0-client-subnet option. Format is an IP address with a subnet mask. Examples: 1.2.3.4/24, 2001:700:300::/48. If you are using DNS-over-HTTPS because of privacy concerns, and do not want any part of your IP address to be sent to authoritative nameservers for geographic location accuracy, use edns_client_subnet=0.0.0.0/0. Google Public DNS normally sends approximate network information (usually replacing the last part of your IPv4 address with zeroes). 0.0.0.0/0 is the default.

Value

data.frame of only answers (use query() for detailed responses)

Note

this is a fairly naive function. It expects Answer to be one of the return value list slots. The intent for it was to make it easier to do bulk forward queries. It will get smarter in future versions.

References

https://developers.google.com/speed/public-dns/docs/dns-over-https

Examples

hosts <- c("rud.is", "dds.ec", "r-project.org", "rstudio.com", "apple.com")
gdns::bulk_query(hosts)

#>             name type  TTL            data
#> 1        rud.is.    1 3599 104.236.112.222
#> 2        dds.ec.    1  299   162.243.111.4
#> 3 r-project.org.    1 7199   137.208.57.37
#> 4   rstudio.com.    1 3599    45.79.156.36
#> 5     apple.com.    1 1006   17.172.224.47
#> 6     apple.com.    1 1006    17.178.96.59
#> 7     apple.com.    1 1006   17.142.160.59
#>